Extract Dag creation, add endpoint for graphviz
